Adding authorization to packages

+++ src/GigManagement/Gig.Application/GigPackages/Commands/DeletePackageCommand.cs
@@ -1,23 +1,28 @@
 ﻿using System;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Hive.Common.Core;
 using Hive.Common.Core.Exceptions;
-using Hive.Common.Core.Security;
+using Hive.Common.Core.Interfaces;
 using Hive.Gig.Application.Interfaces;
 using Hive.Gig.Domain.Entities;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
 
 namespace Hive.Gig.Application.GigPackages.Commands
 {
     public record DeletePackageCommand(int PackageId) : IRequest;
     
-    public class DeletePackageCommandHandler : IRequestHandler<DeletePackageCommand>
+    public class DeletePackageCommandHandler : AuthorizationRequestHandler<Package>, IRequestHandler<DeletePackageCommand>
     {
         private readonly IGigManagementDbContext _context;
         private readonly ILogger<DeletePackageCommand> _logger;
 
-        public DeletePackageCommandHandler(IGigManagementDbContext context, ILogger<DeletePackageCommand> logger)
+        public DeletePackageCommandHandler(IGigManagementDbContext context, 
+            ICurrentUserService currentUserService, IAuthorizationService authorizationService,
+            ILogger<DeletePackageCommand> logger) : base(currentUserService, authorizationService)
         {
             _context = context ?? throw new ArgumentNullException(nameof(context));
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
@@ -31,6 +36,13 @@ namespace Hive.Gig.Application.GigPackages.Commands
             {
                 _logger.LogWarning("Package with id: {@Id} was not found", request.PackageId);
                 throw new NotFoundException(nameof(Package), request.PackageId);
+            }
+            
+            var result = await base.AuthorizeAsync(package,  new [] {"OnlyOwnerPolicy"});
+            
+            if (!result.All(s => s.Succeeded))
+            {
+                throw new ForbiddenAccessException();
             }
 
             _context.Packages.Remove(package);
